requirements.txt
jinja2==3.1.6
requests==2.32.5
typer==0.16.1
python-dotenv==1.1.1